Abstract #run from TestSuite to TestSuiteRunner
Migrate the following methods from `TestSuite` to `TestSuiteRunner`.

- `#run`
- `#run_startup`
- `#run_tests`
- `#run_test`
- `#run_shutdown`
- `#handle_exception`

Then invoke `TestSuiteRunner` from `TestSuite`.

for future parallelization support. Part of GH-235.
